Transform to generalized parser interface
